+++ tests/Test/ConstrainedPlacementByCountTestItem.php
@@ -20,7 +20,7 @@ use function iterator_to_array;
 class ConstrainedPlacementByCountTestItem extends TestItem implements ConstrainedPlacementItem
 {
     /**
-     * @var int
+     * @var float
      */
     public static $limit = 3;
 
@@ -29,12 +29,12 @@ class ConstrainedPlacementByCountTestItem extends TestItem implements Constraine
      */
     public function canBePacked(
         PackedBox $packedBox,
-        int $proposedX,
-        int $proposedY,
-        int $proposedZ,
-        int $width,
-        int $length,
-        int $depth
+        float $proposedX,
+        float $proposedY,
+        float $proposedZ,
+        float $width,
+        float $length,
+        float $depth
     ): bool {
         $alreadyPackedType = array_filter(
             iterator_to_array($packedBox->getItems(), false),
